Option to send Unknown event identify responses.
Adds a flag to the BitProducerConsumer to always respond "Unknown" state to producer / consumer identify requests. This allows creating "non-authoritative" producers and consumers that do not contribute to a distributed state recovery algorithm. That may be helpful for computers that do not have actual hardware that they drive in a particular way.