Adds copy and paste to tabs in the CDI display (#231)
This adds a contextual menu (right click, ctrl click) to tabs in the CDI display that contains "Copy" and "Paste" options. Copy and paste works both within one window and across windows. You can only paste into a tab of the same type from which you copied. The data stored in the clipboard is in the format of the usual backup files.

===

* minor Javadoc

* minor Javadoc

* refactor for generality

* add String LCD method

* Javadoc

* Javadoc

* initial group repl copy and paste

* now with copy from UI

* handle no description tabs

* close intermediate stream

* minor fixes from review, more coming

* refactor output code

* properly handle non-zero-based maps

* use rep.key instead of recreating

* system alert when paste fails

* remove debug printout